+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
   "polars/polars-io",
   "polars/polars-lazy",
   "polars/polars-time",
+  "polars/polars-utils",
 ]
 
 # [patch.crates-io]
