+++ polars/polars-core/src/utils/mod.rs
@@ -21,22 +21,6 @@ impl<T> Deref for Wrap<T> {
     }
 }
 
-unsafe fn index_of_unchecked<T>(slice: &[T], item: &T) -> usize {
-    (item as *const _ as usize - slice.as_ptr() as usize) / std::mem::size_of::<T>()
-}
-
-fn index_of<T>(slice: &[T], item: &T) -> Option<usize> {
-    debug_assert!(std::mem::size_of::<T>() > 0);
-    let ptr = item as *const T;
-    unsafe {
-        if slice.as_ptr() < ptr && slice.as_ptr().add(slice.len()) > ptr {
-            Some(index_of_unchecked(slice, item))
-        } else {
-            None
-        }
-    }
-}
-
 pub(crate) fn set_partition_size() -> usize {
     let mut n_partitions = POOL.current_num_threads();
     // set n_partitions to closes 2^n above the no of threads.
@@ -160,105 +144,6 @@ pub fn slice_offsets(offset: i64, length: usize, array_len: usize) -> (usize, us
         // empty slice
     } else {
         (array_len, 0)
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct Node(pub usize);
-
-impl Default for Node {
-    fn default() -> Self {
-        Node(usize::MAX)
-    }
-}
-
-#[derive(Clone)]
-#[cfg(feature = "private")]
-pub struct Arena<T> {
-    items: Vec<T>,
-}
-
-impl<T> Default for Arena<T> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Simple Arena implementation
-/// Allocates memory and stores item in a Vec. Only deallocates when being dropped itself.
-impl<T> Arena<T> {
-    pub fn add(&mut self, val: T) -> Node {
-        let idx = self.items.len();
-        self.items.push(val);
-        Node(idx)
-    }
-
-    pub fn pop(&mut self) -> Option<T> {
-        self.items.pop()
-    }
-
-    pub fn len(&self) -> usize {
-        self.items.len()
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.items.is_empty()
-    }
-
-    pub fn new() -> Self {
-        Arena { items: vec![] }
-    }
-
-    pub fn with_capacity(cap: usize) -> Self {
-        Arena {
-            items: Vec::with_capacity(cap),
-        }
-    }
-
-    pub fn get_node(&self, val: &T) -> Option<Node> {
-        index_of(&self.items, val).map(Node)
-    }
-
-    #[inline]
-    pub fn get(&self, idx: Node) -> &T {
-        debug_assert!(idx.0 < self.items.len());
-        unsafe { self.items.get_unchecked(idx.0) }
-    }
-
-    #[inline]
-    pub fn get_mut(&mut self, idx: Node) -> &mut T {
-        debug_assert!(idx.0 < self.items.len());
-        unsafe { self.items.get_unchecked_mut(idx.0) }
-    }
-
-    #[inline]
-    pub fn replace(&mut self, idx: Node, val: T) {
-        let x = self.get_mut(idx);
-        *x = val;
-    }
-}
-
-impl<T: Default> Arena<T> {
-    #[inline]
-    pub fn take(&mut self, idx: Node) -> T {
-        std::mem::take(self.get_mut(idx))
-    }
-
-    pub fn replace_with<F>(&mut self, idx: Node, f: F)
-    where
-        F: FnOnce(T) -> T,
-    {
-        let val = self.take(idx);
-        self.replace(idx, f(val));
-    }
-
-    pub fn try_replace_with<F>(&mut self, idx: Node, mut f: F) -> Result<()>
-    where
-        F: FnMut(T) -> Result<T>,
-    {
-        let val = self.take(idx);
-        self.replace(idx, f(val)?);
-        Ok(())
     }
 }
